pyproject.toml
[tool.poetry]
name = "databridgex2"
version = "0.1.0"
description = "Sistema de ETL"
authors = ["Leandro Silva <leandro.silva@escolamobile.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
sqlalchemy = "^2.0.37"
pandas = "^2.2.3"
pymssql = "^2.3.2"
pyarrow = "^19.0.0"
fastparquet = "^2024.11.0"
mysql-connector-python = "^9.2.0"
google-cloud-bigquery = "^3.29.0"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-cloud-core = "^2.4.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
